+++ lib/log/modular_test.go
@@ -65,6 +65,63 @@ func TestModules(t *testing.T) {
 	}
 }
 
+func TestStaticFields(t *testing.T) {
+	loggerConfig := NewConfig()
+	loggerConfig.AddTimeStamp = false
+	loggerConfig.JSONFormat = true
+	loggerConfig.Prefix = "root"
+	loggerConfig.LogLevel = "WARN"
+	loggerConfig.StaticFields = map[string]string{
+		"@service": "benthos_service",
+		"@system":  "foo",
+	}
+
+	buf := LogBuffer{data: ""}
+
+	logger := New(&buf, loggerConfig)
+	logger.Warnln("Warning message root module")
+	logger.Warnf("Warning message root module\n")
+
+	logger2 := logger.NewModule(".foo")
+	logger2.Warnln("Warning message root.foo module")
+
+	expected := `{"@service":"benthos_service","@system":"foo","level":"WARN","component":"root","message":"Warning message root module"}
+{"@service":"benthos_service","@system":"foo","level":"WARN","component":"root","message":"Warning message root module\n"}
+{"@service":"benthos_service","@system":"foo","level":"WARN","component":"root.foo","message":"Warning message root.foo module"}
+`
+
+	if expected != buf.data {
+		t.Errorf("%v != %v", expected, buf.data)
+	}
+}
+
+func TestStaticFieldsEmpty(t *testing.T) {
+	loggerConfig := NewConfig()
+	loggerConfig.AddTimeStamp = false
+	loggerConfig.JSONFormat = true
+	loggerConfig.Prefix = "root"
+	loggerConfig.LogLevel = "WARN"
+	loggerConfig.StaticFields = map[string]string{}
+
+	buf := LogBuffer{data: ""}
+
+	logger := New(&buf, loggerConfig)
+	logger.Warnln("Warning message root module")
+	logger.Warnf("Warning message root module\n")
+
+	logger2 := logger.NewModule(".foo")
+	logger2.Warnln("Warning message root.foo module")
+
+	expected := `{"level":"WARN","component":"root","message":"Warning message root module"}
+{"level":"WARN","component":"root","message":"Warning message root module\n"}
+{"level":"WARN","component":"root.foo","message":"Warning message root.foo module"}
+`
+
+	if expected != buf.data {
+		t.Errorf("%v != %v", expected, buf.data)
+	}
+}
+
 func TestFormattedLogging(t *testing.T) {
 	loggerConfig := NewConfig()
 	loggerConfig.AddTimeStamp = false
@@ -143,51 +200,3 @@ func TestLogLevels(t *testing.T) {
 		}
 	}
 }
-
-/*
-func TestLoggerFanOut(t *testing.T) {
-	var bufMain, buf2, buf3 closeBuf
-
-	conf := NewConfig()
-	conf.AddTimeStamp = false
-	l := New(&bufMain, conf)
-
-	l.Infoln("Foo bar 1")
-
-	l.AddWriter(&buf2)
-	l.Infoln("Foo bar 2")
-
-	l.RemoveWriter(&buf2)
-	l.AddWriter(&buf3)
-	l.Infoln("Foo bar 3")
-
-	l.Close()
-
-	exp := `{"level":"INFO","@service":"benthos","message":"Foo bar 1"}
-{"level":"INFO","@service":"benthos","message":"Foo bar 2"}
-{"level":"INFO","@service":"benthos","message":"Foo bar 3"}` + "\n"
-	if act := bufMain.buf.String(); exp != act {
-		t.Errorf("Wrong logged output: %v != %v", act, exp)
-	}
-
-	exp = `{"level":"INFO","@service":"benthos","message":"Foo bar 2"}` + "\n"
-	if act := buf2.buf.String(); exp != act {
-		t.Errorf("Wrong logged output: %v != %v", act, exp)
-	}
-
-	exp = `{"level":"INFO","@service":"benthos","message":"Foo bar 3"}` + "\n"
-	if act := buf3.buf.String(); exp != act {
-		t.Errorf("Wrong logged output: %v != %v", act, exp)
-	}
-
-	if bufMain.hasClosed {
-		t.Error("bufMain was closed")
-	}
-	if !buf2.hasClosed {
-		t.Error("buf2 not closed")
-	}
-	if !buf3.hasClosed {
-		t.Error("buf3 not closed")
-	}
-}
-*/